Restructured code for better testing
The buienradar requesting and parsing is now split of the maint program so it can be tested without a redefined main-method compile error.
Added a test that does nothing, for now
